setting up testing, renaming utilities to color
