+++ Makefile
@@ -1,0 +1,9 @@
+.PHONY: check
+
+test:
+	pytest --cov=great_tables --cov-report=xml
+
+test-update:
+	pytest --snapshot-update
+
+
