+++ gt_extras/__init__.py
@@ -8,8 +8,8 @@ from .themes import gt_theme_dot_matrix
 from .themes import gt_theme_dark
 from .themes import gt_theme_pff
 
-from .utilities import gt_highlight_cols
-from .utilities import gt_hulk_col_numeric
+from .colors import gt_highlight_cols
+from .colors import gt_hulk_col_numeric
 
 __all__ = [
     "gt_theme_538",
